remove agora-rtc-sdk-ng package to provide consistent use across all components to use agora-rtc-react
