+++ client/package.json
@@ -22,7 +22,6 @@
     "agora-extension-spatial-audio": "^1.0.0-beta1",
     "agora-extension-virtual-background": "^1.2.0",
     "agora-rtc-react": "^2.1.0",
-    "agora-rtc-sdk-ng": "^4.20.0",
     "gsap": "^3.12.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
