+++ client/src/AgoraManager/agora.tsx
@@ -1,17 +1,14 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import AgoraRTC from 'agora-rtc-sdk-ng';
-import { AgoraRTCProvider } from 'agora-rtc-react';
-import { IAgoraRTCClient as IAgoraRTCClientSDK } from 'agora-rtc-sdk-ng';
-import { IAgoraRTCClient as IAgoraRTCClientReact } from 'agora-rtc-react';
+import AgoraRTC, { AgoraRTCProvider, IAgoraRTCClient } from 'agora-rtc-react';
 
 interface ClientProps {
     children?: React.ReactNode;
 }
 
 export const Client: React.FC<ClientProps> = ({ children }) => {
-    const client: IAgoraRTCClientSDK = AgoraRTC.createClient({ mode: 'rtc', codec: 'vp8' });
-    return <AgoraRTCProvider client={client as unknown as IAgoraRTCClientReact}>{children}</AgoraRTCProvider>;
+    const client: IAgoraRTCClient = AgoraRTC.createClient({ mode: 'rtc', codec: 'vp8' }) as unknown as IAgoraRTCClient;
+    return <AgoraRTCProvider client={client}>{children}</AgoraRTCProvider>;
 };
 
 const container = document.getElementById('container');
