+++ client/src/store/user.ts
@@ -1,17 +1,6 @@
 import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
 import { User } from '../interfaces/user';
 
-export interface UserForm {
-    id: string;
-    username: string;
-    picUrl: File;
-    about: string;
-    github: string;
-    linkedin: string;
-    portfolio: string;
-    leetcode: string;
-}
-
 export const getUser = createAsyncThunk<User | null, number, { rejectValue: {} | string }>(
     'user/getUser',
     async (id, {rejectWithValue}) => {
